requirements.txt
torch>=1.3
gym[mujoco]>=0.15
tqdm>=4.0
pyyaml>=5.1
